serp_backend/requirements.txt

python-dotenv
serpapi
flask
flask-cors
